requirements.txt
curl_cffi>=0.5.10
playwright>=1.40.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
better-sqlite3-utils>=0.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
asyncio>=3.4.3
